refactor: UserRecordModal 위치 변경
